+++ inlang/source-code/ide-extension/src/diagnostics/linterDiagnostics.ts
@@ -1,8 +1,13 @@
 import * as vscode from "vscode"
 import { state } from "../utilities/state.js"
 import type { MessageLintReport } from "@inlang/sdk"
+import { recommendNinja } from "../utilities/recommend/ninja/ninja.js"
+import type { NodeishFilesystem } from "@lix-js/fs"
 
-export async function linterDiagnostics(args: { context: vscode.ExtensionContext }) {
+export async function linterDiagnostics(args: {
+	context: vscode.ExtensionContext
+	nodeishFs: NodeishFilesystem
+}) {
 	const linterDiagnosticCollection = vscode.languages.createDiagnosticCollection("inlang-lint")
 
 	async function updateLintDiagnostics() {
@@ -82,6 +87,10 @@ export async function linterDiagnostics(args: { context: vscode.ExtensionContext
 								activeTextEditor.document.uri,
 								flattenDiagnostics(diagnosticsIndex)
 							)
+
+							if (diagnostics.length > 0) {
+								recommendNinja({ fs: args.nodeishFs })
+							}
 						}
 					)
 				}